requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
jinja2==3.1.6
python-multipart==0.0.20
passlib==1.7.4
requests==2.32.5
itsdangerous==2.2.0
httpx==0.28.1
pyyaml==6.0.3
